+++ CHANGES.md
@@ -1,3 +1,7 @@
+## vX.X.X (XXXX-XX-XX)
+
+- Add `from_filename()` function.
+
 ## v0.0.1 (2014-07-15)
 
 - First release